Refactorize cfe values and add section to dashboard
