+++ src/app/cfe/CfeValues.vue
@@ -33,82 +33,9 @@
                             <b-col
                                 cols="12"
                                 class="prices">
-                                <b-form
-                                    class="prices-form"
-                                    @submit="(e) => {e.preventDefault()}">
-                                    <div class="edit-buttons text-right">
-                                        <span
-                                            @click="savePrices"
-                                            v-if="isEditing">
-                                            <i class="far fa-save"></i>
-                                        </span>
-                                        <span
-                                            @click="cancelEditing"
-                                            v-if="isEditing">
-                                            <i class="fas fa-trash-alt"></i>
-                                        </span>
-                                        <span
-                                            @click="editFields"
-                                            v-else>
-                                            <i class="fas fa-edit"></i>
-                                        </span>
-                                    </div>
-                                    <b-form-group
-                                        label="Precio base:"
-                                        label-for="base_price">
-                                        <b-form-input
-                                            v-model="base"
-                                            id="base_price"
-                                            :disabled="!isEditing"
-                                            step="any"
-                                            type="number">
-                                        </b-form-input>
-                                    </b-form-group>
-                                    <b-form-group
-                                        label="Precio media:"
-                                        label-for="middle_price">
-                                        <b-form-input
-                                            v-model="middle"
-                                            id="middle_price"
-                                            :disabled="!isEditing"
-                                            step="any"
-                                            type="number">
-                                        </b-form-input>
-                                    </b-form-group>
-                                    <b-form-group
-                                        label="Precio punta:"
-                                        label-for="peak_price">
-                                        <b-form-input
-                                            v-model="peak"
-                                            id="peak_price"
-                                            :disabled="!isEditing"
-                                            step="any"
-                                            type="number">
-                                        </b-form-input>
-                                    </b-form-group>
-                                    <b-form-group
-                                        label="Precio capacidad:"
-                                        label-for="capacity_price">
-                                        <b-form-input
-                                            v-model="capacity"
-                                            id="capacity_price"
-                                            :disabled="!isEditing"
-                                            step="any"
-                                            type="number">
-                                        </b-form-input>
-                                    </b-form-group>
-                                    <b-form-group
-                                        label="Precio distribución:"
-                                        label-for="distribution_price">
-                                        <b-form-input
-                                            v-model="distribution"
-                                            id="distribution_price"
-                                            :disabled="!isEditing"
-                                            step="any"
-                                            type="number">
-                                        </b-form-input>
-                                    </b-form-group>
-                                </b-form>
+                                <v-cfe
+                                    :fullWidth="false"
+                                    :forceCurrentMonth="false"/>
                             </b-col>
                         </b-row>
                     </b-col>
@@ -120,23 +47,18 @@
 
 <script>
 import moment from 'moment';
+import VCfe from '@/app/components/VCfe';
 import _l from 'lodash';
 
 export default {
-    data() {
-        return {
-            base: null,
-            middle: null,
-            peak: null,
-            capacity: null,
-            distribution: null,
-            isEditing: false
-        }
+    components: {
+        VCfe
     },
 
-    beforeMount() {
-        this.resetPrices();
-        this.$store.dispatch('meter/changeCfeperiod', {years: 0, months: 0});
+    data() {
+        return {
+            
+        }
     },
 
     computed: {
@@ -155,103 +77,16 @@ export default {
             },
 
             set() {}
-        },
-
-        basePrice() {
-            return this.$store.state.meter.cfeValues.prices.base;
-        },
-        middlePrice() {
-            return this.$store.state.meter.cfeValues.prices.middle;
-        },
-        peakPrice() {
-            return this.$store.state.meter.cfeValues.prices.peak;
-        },
-        capacityPrice() {
-            return this.$store.state.meter.cfeValues.prices.capacity;
-        },
-        distributionPrice() {
-            return this.$store.state.meter.cfeValues.prices.distribution;
-        }
-    },
-
-    watch: {
-        basePrice(newValue) {
-            this.base = newValue;
-        },
-        middlePrice(newValue) {
-            this.middle = newValue;
-        },
-        peakPrice(newValue) {
-            this.peak = newValue;
-        },
-        capacityPrice(newValue) {
-            this.capacity = newValue;
-        },
-        distributionPrice(newValue) {
-            this.distribution = newValue;
         }
     },
 
     methods: {
-        savePrices() {
-            if (parseFloat(this.base) > 0 &&
-                parseFloat(this.middle) > 0 &&
-                parseFloat(this.peak) > 0 &&
-                parseFloat(this.capacity) > 0 &&
-                parseFloat(this.distribution) > 0) {
-                let data = {
-                    base: this.base,
-                    middle: this.middle,
-                    peak: this.peak,
-                    capacity: this.capacity,
-                    distribution: this.distribution
-                }
-                this.$store.dispatch('meter/setCfePrices', data)
-                    .then(() => {
-                        this.$notify({
-                            group: 'notification',
-                            type: 'success',
-                            title: 'Los precios de este mes se cambiaron correctamente'
-                        });
-                    })
-                    .catch(err => {
-                        this.$notify({
-                            group: 'notification',
-                            type: 'error',
-                            title: 'No fue posible cambiar los precios'
-                        });
-                    })
-                    .finally(() => {
-                        this.isEditing = false;
-                    });
-            } else {
-                this.$notify({
-                            group: 'notification',
-                            type: 'warn',
-                            title: 'Verifica que los campos estén llenados de forma correcta'
-                        });
-            }
-        },
-        resetPrices() {
-            this.base = this.basePrice;
-            this.middle = this.middlePrice;
-            this.peak = this.peakPrice;
-            this.capacity = this.capacityPrice;
-            this.distribution = this.distributionPrice;
-        },
-        editFields() {
-            this.isEditing = true;
-        },
-        cancelEditing() {
-            this.isEditing = false;
-            this.resetPrices();
-        },
         changePeriod(type, quantity) {
             if (!this.isEditing) {
                 if (type === 0) {   // Year
-                    this.$store.dispatch('meter/changeCfeperiod', {years: quantity, months: 0});
+                    this.$store.dispatch('meter/changeCfePeriod', {years: quantity, months: 0});
                 } else if (type === 1) {   // Month
-                    this.$store.dispatch('meter/changeCfeperiod', {years: 0, months: quantity});
+                    this.$store.dispatch('meter/changeCfePeriod', {years: 0, months: quantity});
                 }
             }
         }
@@ -300,28 +135,9 @@ $darkgray: #485658;
                     }
                 }
             }
-        }
 
-        .prices {
-            margin: 4rem 0;
-
-            .edit-buttons {
-                margin-bottom: 2rem;
-
-                span {
-                    &:hover {
-                        cursor: pointer;
-                    }
-                    margin-right: 1rem;
-                    &:last-child {
-                        margin-right: 0;
-                    }
-                }
-            }
-
-            .prices-form {
-                width: 30% !important;
-                margin: 0 auto;
+            .prices {
+                margin: 4rem 0;
             }
         }
     }
