+++ src/store/modules/meter/mutations-types.js
@@ -10,3 +10,4 @@ export const DELETE_ASSIGNED = 'deleteMeterAssigned';
 export const UPDATE_ASSIGNED = 'updateMeterAssigned';
 export const DELETE_ALL_ASSIGNED = 'deleteAllMeterAssigned';
 export const GET_CFE_VALUES = 'getCfeValues';
+export const GET_CURRENT_CFE_VALUES = 'getCurrentCfeValues';
